Fork oslo.config.generator from stable/icehouse
Because we'll need to generate configuration files from external
libraries, we need to fork the generator from oslo.  So, for now
we just did a direct copy from the stable/icehouse branch and
removed the module from openstack-common.conf.

Once we update our libraries to stable/icehouse we can basically
revert this commit.

Change-Id: I2c9b754beca237c775c5ac4967bbc8fdbc49df00
Signed-off-by: Paul Belanger <paul.belanger@polybeacon.com>